util: is_uefi_bootable() does not imply EFI variables exist
A system may be booted via UEFI but not provide EFI variables at runtime.

In this case directory /sys/firmware/efi exists
but directory /sys/firmware/efi/efivars does not exist.

We should not try to mount the non-existent directory.
This implies that we should not add an efibootmgr command in
gen_uefi_install_commands().

Suggested-by: Dan Bungert <daniel.bungert@canonical.com>
Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>